Refactor protocol list footer
- Moved fallback chain icons to new networks file under ui/utils
- Renamed CustomRPCFooter to TopMenuProtocolListFooter and moved it to a separate file
- Added new utility fn to get network icon name
- Fix color in new settings icon svg
- Fixed an issue with the test networks label decoration z-index
- Fixed excess spacing around the footer
