+++ ui/components/Shared/SharedNetworkIcon.tsx
@@ -1,5 +1,6 @@
 import { EVMNetwork } from "@tallyho/tally-background/networks"
 import React, { ReactElement } from "react"
+import { getNetworkIconName } from "../../utils/networks"
 
 export default function SharedNetworkIcon(props: {
   network: EVMNetwork
@@ -8,6 +9,8 @@ export default function SharedNetworkIcon(props: {
   backgroundOpacity?: number
 }): ReactElement {
   const { network, size, hasBackground = false, backgroundOpacity = 1 } = props
+  const iconName = getNetworkIconName(network)
+
   return (
     <>
       <div className="icon_network_wrapper">
@@ -34,9 +37,7 @@ export default function SharedNetworkIcon(props: {
           opacity: ${backgroundOpacity};
         }
         .icon_network {
-          background: url("./images/networks/${network.name
-            ?.replaceAll(" ", "")
-            .toLowerCase()}-square@2x.png");
+          background: url("./images/networks/${iconName}-square@2x.png");
           background-size: cover;
           height: ${size - 6}px;
           width: ${size - 6}px;
